+++ scripts/test/Selenium/Agent/AgentAppointmentNotification.t
@@ -650,10 +650,10 @@ $Selenium->RunTest(
             # custom datetime notification $YearBeforeLast-09-01 10:10:00
             {
                 Data => {
-                    Description                           => "Custom datetime $YearBeforeLastSettings->{Year}-09-01 10:10:00",
-                    NotificationTemplate                  => 'Custom',
-                    NotificationCustomRelativeInput       => 0,
-                    NotificationCustomDateTimeInput       => 1,
+                    Description                     => "Custom datetime $YearBeforeLastSettings->{Year}-09-01 10:10:00",
+                    NotificationTemplate            => 'Custom',
+                    NotificationCustomRelativeInput => 0,
+                    NotificationCustomDateTimeInput => 1,
                     NotificationCustomRelativeUnitCount   => 0,
                     NotificationCustomRelativeUnit        => 'minutes',
                     NotificationCustomRelativePointOfTime => 'beforestart',
@@ -678,10 +678,10 @@ $Selenium->RunTest(
             # custom datetime notification $YearBeforeLast-10-18 00:03:00
             {
                 Data => {
-                    Description                           => "Custom datetime $YearBeforeLastSettings->{Year}-10-18 01:03:00",
-                    NotificationTemplate                  => 'Custom',
-                    NotificationCustomRelativeInput       => 0,
-                    NotificationCustomDateTimeInput       => 1,
+                    Description                     => "Custom datetime $YearBeforeLastSettings->{Year}-10-18 01:03:00",
+                    NotificationTemplate            => 'Custom',
+                    NotificationCustomRelativeInput => 0,
+                    NotificationCustomDateTimeInput => 1,
                     NotificationCustomRelativeUnitCount   => 0,
                     NotificationCustomRelativeUnit        => 'minutes',
                     NotificationCustomRelativePointOfTime => 'beforestart',
@@ -734,10 +734,10 @@ $Selenium->RunTest(
             # custom datetime notification $YearBeforeLast-10-18 02:03:00
             {
                 Data => {
-                    Description                           => "Custom datetime $YearBeforeLastSettings->{Year}-10-18 02:03:00",
-                    NotificationTemplate                  => 'Custom',
-                    NotificationCustomRelativeInput       => 0,
-                    NotificationCustomDateTimeInput       => 1,
+                    Description                     => "Custom datetime $YearBeforeLastSettings->{Year}-10-18 02:03:00",
+                    NotificationTemplate            => 'Custom',
+                    NotificationCustomRelativeInput => 0,
+                    NotificationCustomDateTimeInput => 1,
                     NotificationCustomRelativeUnitCount   => 0,
                     NotificationCustomRelativeUnit        => 'minutes',
                     NotificationCustomRelativePointOfTime => 'beforestart',